requirements.txt
fastapi
passlib
pydantic==1.10.11
python-jose
uvicorn
asyncpg
bcrypt
pytest
httpx
pytest-mock
pytest-asyncio